+++ src/game/BridgeData.java
@@ -4,7 +4,7 @@ public class BridgeData {
 	int count;
 	TileIndex start_tile;
 	TileIndex end_tile;
-	byte type;
+	int type;
 	byte [] indexes = new byte[Bridge.MAX_BRIDGES];
 	int  [] costs  = new int[Bridge.MAX_BRIDGES];
