+++ src/game/EngineInfo.java
@@ -6,11 +6,11 @@ package game;
 
 public class EngineInfo {
 	public int base_intro;
-	public byte unk2;              ///< Carriages have the highest bit set in this one
-	public byte lifelength;
-	public byte base_life;
-	public byte railtype;
-	public byte climates;
+	public int unk2;              ///< Carriages have the highest bit set in this one
+	public int lifelength;
+	public int base_life;
+	public int railtype;
+	public int climates;
 	public int refit_mask;
 
 }
